k8s-multi-app/pyproject.toml
[project]
name = "k8s-multi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.7",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
